use QueryValue instead of Mixed in queries
